+++ client/slices/address/types/config.ts
@@ -1,5 +1,3 @@
-import type { ArrayElement } from 'types/utils';
-
 export const IDENTICON_TYPES = [
   'github',
   'jazzicon',
@@ -8,13 +6,13 @@ export const IDENTICON_TYPES = [
   'nouns',
 ] as const;
 
-export type IdenticonType = ArrayElement<typeof IDENTICON_TYPES>;
+export type IdenticonType = typeof IDENTICON_TYPES[number];
 
 export const ADDRESS_VIEWS_IDS = [
   'top_accounts',
 ] as const;
 
-export type AddressViewId = ArrayElement<typeof ADDRESS_VIEWS_IDS>;
+export type AddressViewId = typeof ADDRESS_VIEWS_IDS[number];
 
 export const ADDRESS_FORMATS = [ 'base16', 'bech32' ] as const;
 export type AddressFormat = typeof ADDRESS_FORMATS[ number ];
