+++ client/slices/contract/types/client.ts
@@ -1,11 +1,5 @@
 import type { SmartContractLicenseType } from './api';
 
-export interface ContractCodeIde {
-  title: string;
-  url: string;
-  icon_url: string;
-}
-
 export interface ContractLicense {
   type: SmartContractLicenseType;
   url: string;
