+++ client/slices/home/hooks/useChartDataQuery.ts
@@ -1,3 +1,4 @@
+import type { ChainIndicatorId } from 'client/slices/home/types/config';
 import type { LineChartData } from 'toolkit/components/charts/line/types';
 
 import useApiQuery from 'client/api/hooks/useApiQuery';
@@ -5,7 +6,6 @@ import useApiQuery from 'client/api/hooks/useApiQuery';
 import { getChartData } from 'client/slices/home/utils/chart';
 
 import config from 'configs/app';
-import type { ChainIndicatorId } from 'configs/app/ui/homepage';
 
 const rollupFeature = config.features.rollup;
 const isOptimisticRollup = rollupFeature.isEnabled && rollupFeature.type === 'optimistic';
