+++ configs/app/ui.ts
@@ -1,4 +1,4 @@
-import type { ContractCodeIde } from 'client/slices/contract/types/client';
+import type { ContractCodeIde } from 'client/slices/contract/types/config';
 import { type NavItemExternal, type NavigationLayout, type NavigationPromoBannerConfig } from 'types/client/navigation';
 import type { NetworkExplorer } from 'types/networks';
 import type { ColorThemeId } from 'types/settings';
