+++ configs/app/ui/homepage.ts
@@ -1,48 +1,14 @@
-export const CHAIN_INDICATOR_IDS = [ 'daily_txs', 'daily_operational_txs', 'coin_price', 'secondary_coin_price', 'market_cap', 'tvl' ] as const;
-export type ChainIndicatorId = typeof CHAIN_INDICATOR_IDS[number];
+import type { ChainIndicatorId, HeroBannerConfig, HomeStatsWidgetId } from 'client/slices/home/types/config';
+import { HOME_STATS_WIDGET_IDS } from 'client/slices/home/types/config';
 
-export const HOME_STATS_WIDGET_IDS = [
-  'latest_batch',
-  'total_blocks',
-  'average_block_time',
-  'total_txs',
-  'total_operational_txs',
-  'latest_l1_state_batch',
-  'wallet_addresses',
-  'gas_tracker',
-  'btc_locked',
-  'current_epoch',
-] as const;
-export type HomeStatsWidgetId = typeof HOME_STATS_WIDGET_IDS[number];
-
-export interface HeroBannerButtonState {
-  background?: Array<string | undefined>;
-  text_color?: Array<string | undefined>;
-}
-
-export interface HeroBannerConfig {
-  background?: Array<string | undefined>;
-  text_color?: Array<string | undefined>;
-  border?: Array<string | undefined>;
-  button?: {
-    _default?: HeroBannerButtonState;
-    _hover?: HeroBannerButtonState;
-    _selected?: HeroBannerButtonState;
-  };
-  search?: {
-    border_width?: Array<string | undefined>;
-  };
-  text?: string;
-}
-
-import * as features from 'configs/app/features';
+import { rollup } from 'configs/app/features';
 import { getEnvValue, getExternalAssetFilePath, parseEnvJson } from 'configs/app/utils';
 
 const homePageStats: Array<HomeStatsWidgetId> = (() => {
   const parsedValue = parseEnvJson<Array<HomeStatsWidgetId>>(getEnvValue('NEXT_PUBLIC_HOMEPAGE_STATS'));
 
   if (!Array.isArray(parsedValue)) {
-    const rollupFeature = features.rollup;
+    const rollupFeature = rollup;
 
     if (rollupFeature.isEnabled && [ 'zkSync', 'arbitrum' ].includes(rollupFeature.type)) {
       return [ 'latest_batch', 'average_block_time', 'total_txs', 'wallet_addresses', 'gas_tracker' ];
