+++ configs/app/ui/views/address.ts
@@ -1,8 +1,7 @@
-import type { AddressFormat, AddressViewId, IdenticonType } from 'client/slices/address/types/view';
-import { ADDRESS_FORMATS, ADDRESS_VIEWS_IDS, IDENTICON_TYPES } from 'client/slices/address/types/view';
-
-import type { SmartContractVerificationMethodExtra } from 'client/slices/contract/pages/contract-verification/utils';
-import { SMART_CONTRACT_EXTRA_VERIFICATION_METHODS } from 'client/slices/contract/pages/contract-verification/utils';
+import type { AddressFormat, AddressViewId, IdenticonType } from 'client/slices/address/types/config';
+import { ADDRESS_FORMATS, ADDRESS_VIEWS_IDS, IDENTICON_TYPES } from 'client/slices/address/types/config';
+import type { SmartContractVerificationMethodExtra } from 'client/slices/contract/types/config';
+import { SMART_CONTRACT_EXTRA_VERIFICATION_METHODS } from 'client/slices/contract/types/config';
 
 import { getEnvValue, parseEnvJson } from 'configs/app/utils';
