+++ lib/settings/identIcon.ts
@@ -1,4 +1,4 @@
-import type { IdenticonType } from 'client/slices/address/types/view';
+import type { IdenticonType } from 'client/slices/address/types/config';
 
 export const IDENTICONS: Array<{ label: string; id: IdenticonType; sampleBg: string }> = [
   {
